refactor(bibe): move the BIBE-PDU codec into its own module
